@concave/core - remove circular deps & fix bundle
