+++ libs/core/package.json
@@ -3,17 +3,36 @@
   "version": "0.0.0",
   "description": "Concave Core library",
   "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ConcaveFi/concave-frontend.git",
+    "directory": "libs/core"
+  },
   "main": "./dist/index.js",
   "module": "./dist/index.mjs",
   "types": "./dist/index.d.ts",
   "sideEffects": false,
   "files": [
-    "dist/**"
+    "dist"
   ],
   "scripts": {
-    "build": "tsup src/index.ts --format esm,cjs --dts",
-    "dev": "tsup src/index.ts --format esm,cjs --watch --dts",
+    "build": "tsup",
+    "dev": "tsup --watch",
     "test": "tsdx test --passWithNoTests"
+  },
+  "tsup": {
+    "entry": [
+      "src/**"
+    ],
+    "format": [
+      "esm"
+    ],
+    "dts": true,
+    "splitting": true,
+    "treeshake": true,
+    "sourcemap": true,
+    "clean": true,
+    "silent": true
   },
   "dependencies": {
     "@ethersproject/address": "^5.6.1",
