--- libs/core/src/declarations.d.ts
@@ -1,1 +1,0 @@
-declare module 'toformat'
