+++ libs/core/src/entities/currency.ts
@@ -1,4 +1,146 @@
-import { NativeCurrency } from './nativeCurrency'
-import { Token } from './token'
+import invariant from 'tiny-invariant'
+import { validateAndParseAddress } from '../functions/validateAndParseAddress'
+
+/**
+ * A currency is any fungible financial instrument, including Ether, all ERC20 tokens, and other chain-native currencies
+ */
+export abstract class AbstractCurrency {
+  /**
+   * Returns whether the currency is native to the chain and must be wrapped (e.g. Ether)
+   */
+  public abstract readonly isNative: boolean
+  /**
+   * Returns whether the currency is a token that is usable in Uniswap without wrapping
+   */
+  public abstract readonly isToken: boolean
+
+  /**
+   * The chain ID on which this currency resides
+   */
+  public readonly chainId: number
+  /**
+   * The decimals used in representing currency amounts
+   */
+  public readonly decimals: number
+  /**
+   * The symbol of the currency, i.e. a short textual non-unique identifier
+   */
+  public readonly symbol?: string
+  /**
+   * The name of the currency, i.e. a descriptive textual non-unique identifier
+   */
+  public readonly name?: string
+  address: string
+
+  /**
+   * Constructs an instance of the base class `BaseCurrency`.
+   * @param chainId the chain ID on which this currency resides
+   * @param decimals decimals of the currency
+   * @param symbol symbol of the currency
+   * @param name of the currency
+   */
+  protected constructor(chainId: number, decimals: number, symbol?: string, name?: string) {
+    invariant(Number.isSafeInteger(chainId), 'CHAIN_ID')
+    invariant(decimals >= 0 && decimals < 255 && Number.isInteger(decimals), 'DECIMALS')
+
+    this.chainId = chainId
+    this.decimals = decimals
+    this.symbol = symbol
+    this.name = name
+  }
+
+  /**
+   * Returns whether this currency is functionally equivalent to the other currency
+   * @param other the other currency
+   */
+  public abstract equals(other: AbstractCurrency): boolean
+
+  /**
+   * Return the wrapped version of this currency that can be used with the Uniswap contracts. Currencies must
+   * implement this to be used in Uniswap
+   */
+  public abstract get wrapped(): Token
+}
+
+/**
+ * Represents an ERC20 token with a unique address and some metadata.
+ */
+export class Token extends AbstractCurrency {
+  public readonly isNative: false = false
+  public readonly isToken: true = true
+
+  public declare readonly address: string
+
+  public constructor(
+    chainId: number,
+    address: string,
+    decimals: number,
+    symbol?: string,
+    name?: string,
+  ) {
+    super(chainId, decimals, symbol, name)
+    this.address = validateAndParseAddress(address)
+  }
+
+  /**
+   * Returns true if the two tokens are equivalent, i.e. have the same chainId and address.
+   * @param other other token to compare
+   */
+  public equals(other: AbstractCurrency): boolean {
+    return (
+      other && other.isToken && this.chainId === other.chainId && this.address === other.address
+    )
+  }
+
+  /**
+   * Returns true if the address of this token sorts before the address of the other token
+   * @param other other token to compare
+   * @throws if the tokens have the same address
+   * @throws if the tokens are on different chains
+   */
+  public sortsBefore(other: Token): boolean {
+    invariant(this.chainId === other.chainId, 'CHAIN_IDS')
+    invariant(this.address !== other.address, 'ADDRESSES')
+    return this.address.toLowerCase() < other.address.toLowerCase()
+  }
+
+  /**
+   * Return this token, which does not need to be wrapped
+   */
+  public get wrapped(): Token {
+    return this
+  }
+}
+
+/**
+ * Compares two currencies for equality
+ */
+export function currencyEquals(currencyA: AbstractCurrency, currencyB: AbstractCurrency): boolean {
+  if (currencyA instanceof Token && currencyB instanceof Token) {
+    return currencyA.equals(currencyB)
+  } else if (currencyA instanceof Token) {
+    return false
+  } else if (currencyB instanceof Token) {
+    return false
+  } else {
+    return currencyA === currencyB
+  }
+}
+
+/**
+ * Represents the native currency of the chain on which it resides, e.g.
+ */
+export abstract class NativeCurrency extends AbstractCurrency {
+  public readonly isNative: true = true
+  public readonly isToken: false = false
+
+  /**
+   * Returns true if the two currencies are equivalent, i.e. both native with same chainId.
+   * @param other other currency to compare
+   */
+  public equals(other: AbstractCurrency): boolean {
+    return other.isNative && this.chainId === other.chainId
+  }
+}
 
 export type Currency = NativeCurrency | Token
