+++ libs/core/src/entities/index.ts
@@ -1,9 +1,6 @@
-export * from './AbstractCurrency'
-// TODO: This is actually a type...
+// organize-imports-ignore
 export * from './currency'
 export * from './currencyAmount'
 export * from './fraction'
 export * from './Native'
-export * from './nativeCurrency'
 export * from './percent'
-export * from './token'
